+++ src/main/java/Listener/ConsoleListener.java
@@ -5,16 +5,10 @@
  */
 package Listener;
 
-import Audio.FM;
+import Audio.Radio;
 import Main.Main;
-import Utility.WebScraper;
-import com.mashape.unirest.http.exceptions.UnirestException;
 import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
 import java.util.Scanner;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import net.dv8tion.jda.core.OnlineStatus;
 
 /**